pyproject.toml
[project]
name = "auth-server"
version = "0.1.0"
description = "Multi-tenant OAuth2 auth server prototype"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.8",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
